Thu Aug 11 16:37:58 2011                        Michael Jennings (mej)
Per discussion with Greg, this version of the parser requires a comma
between items in a list.  Whitespace surrounding the comma is ignored,
but embedded whitespace within a value is significant.  This maintains
consistency between quoting requirements for keys and values but will
require some existing configuration files to be modified.

The test config file has been modified so that all unit tests pass
again.
----------------------------------------------------------------------


git-svn-id: https://warewulf.lbl.gov/svn/trunk@577 3fa55bd7-8392-0410-b022-a3fc164e7bfc