Fri Feb 10 11:12:04 2012                        Michael Jennings (mej)
Added a unit test to flag the "bug" I mentioned in yesterday's commit,
and it turns out that it doesn't adversely affect any desired/correct
behavior.  So rather than do a big revamp, we've decided to leave the
unit tests in place and will maintain the current (technically wrong
but behaviorally correct) code until such time as the unit tests fail
or other problems arise which require us to revisit.
----------------------------------------------------------------------


git-svn-id: https://warewulf.lbl.gov/svn/trunk@752 3fa55bd7-8392-0410-b022-a3fc164e7bfc