common/share/setup.sql   * Change INT id fields to be UNSIGNED
Now you can have double the entries for objects!

http://dev.mysql.com/doc/refman/5.0/en/integer-types.html

It should be noted that trying to insert an entry that is
negative will not result in a failure, but it'll act as
if the field wasn't specified and auto_increment normally.

Trying to go *above* the limit, will drop it down to the
max value, and if you try that more than once, then you'll
get a duplicate key error. At least on MySQL 5.0.84.



git-svn-id: https://warewulf.lbl.gov/svn/trunk@841 3fa55bd7-8392-0410-b022-a3fc164e7bfc