Workaround for bug where mkfs.ext3/ext4 requires /etc/mtab
Filesystem creation with an ext3 or ext4 filesystem fails when
mkfs.ext3/ext4 is run and there is no /etc/mtab file (as in the
bootstrap BusyBox environment). This is a known recent bug
in e2fsprogs:

http://sourceforge.net/tracker/index.php?func=detail&aid=3509398&group_id=2406&atid=102406

And the resulting corrupt filesystem can cause stateful provisioning
to fail.

This may be fixed in the most recent e2fsprogs (hard to tell,
would be v1.42.3) but for the moment I'm taking the easy way out,
and symlinking the /etc/mtab to the /tmp/mtab which is present in
the bootstrap. :-) This allows stateful provisioning to succeed in
my testing.

git-svn-id: https://warewulf.lbl.gov/svn/trunk@987 3fa55bd7-8392-0410-b022-a3fc164e7bfc