20-filesystems -- Adds in the ability to use 'fill' for a size. Be sure it's the *last* entry in the FILESYSTEMS variable, or it won't make you too happy. ;)
    80-mkbootable -- So, I ran into an issue with 'grub-install' (SL
6.3)... it just didn't work. *BUT* running the commands to setup grub
manually did work. So, this is a hack to work around that. If
grub-install fails, it generates a small set of commands, that is ran
in the chroot, which is used as input to grub to setup the boot loader
that way. Any output is sent to: /root/grubinstall.out  -- in
$NEWROOT.



git-svn-id: https://warewulf.lbl.gov/svn/trunk@1280 3fa55bd7-8392-0410-b022-a3fc164e7bfc