Mon Mar 11 14:47:19 2013                        Michael Jennings (mej)
Add DMI data gatherer and corresponding checks.  Definitely not
speedy, but there's an awful lot of very valuable information that can
be gleaned from it.  And, as always, once you take the initial hit,
you can write as many DMI checks as you like with minimal added
overhead.

Still need to write the unit tests for the new checks.
----------------------------------------------------------------------


git-svn-id: https://warewulf.lbl.gov/svn/trunk@1298 3fa55bd7-8392-0410-b022-a3fc164e7bfc