Wed Apr 24 12:53:59-0500 2013                       Jason Stover (jms)
Update: mic/README

There's an issue with the assignment of IP addresses on MIC devices
after an IP address has been changed within the Warewulf Datastore.

This is caused by how the _IPADDR array is handled on the node object,
and, currently, is tagged as a "Known Issue" for the warewulf-mic
package.
----------------------------------------------------------------------


git-svn-id: https://warewulf.lbl.gov/svn/trunk@1367 3fa55bd7-8392-0410-b022-a3fc164e7bfc