Fri Jul 26 16:12:01-0500 2013                       Jason Stover (jms)
Added 'wwshd' script to common/bin/.

This is an example only. It creates a pipe, /tmp/wwshd, that commands
can be echoed into and it will be ran as the user running wwshd.

  $ echo "provision set --vnfs=sl6.2 --bootstrap=sl6.2 node002" > \
    /tmp/wwshd

The pipe is attempted to be chowned root:warewulf, with permissions
of: 0660

Further extensions would be a check on the command comming in to see
if it matches 'delete', or any other destructive command. No output
redirection, so run something like:

  # nohup ./wwshd &

Can be horribly dangerous, might blow up, may even lead to the 
deletion of all your data and make you sad; So this is *only* in SVN
and not added to the distribution.

-J
----------------------------------------------------------------------  


git-svn-id: https://warewulf.lbl.gov/svn/trunk@1407 3fa55bd7-8392-0410-b022-a3fc164e7bfc