Tue Aug 13 23:42:36-0500 2013                       Jason Stover (jms)
Change logic a bit in cgi-bin/vnfs.pl to not exit if a lock cannot be
obtained when generating the cache file. Just return the HTTP 500 
code. The rest of the blocks should be ignored.

May need to add in a $is_lock check in the conditionals, to possibly 
not return text again in the same run.
----------------------------------------------------------------------


git-svn-id: https://warewulf.lbl.gov/svn/trunk@1422 3fa55bd7-8392-0410-b022-a3fc164e7bfc