Mon Nov 18 10:52:20-0600 2013                       Jason Stover (jms)
Add in Object->disable(). This allows a object to be disabled from
normal Warewulf processing. Node handling added in currently.

Example:
  # wwsh node set -D eth0 --hwaddr=00:11:22:33:44:55 comp0001
  # wwsh node set --disable=1 comp0001
  # wwsh node set -D eth0 --hwaddr=00:11:22:33:44:55 newnode

When disbled, the node is written into dynamic_hosts but a PXE file
is not written out. Any existing PXE file is not deleted. Checks
in the cgi-bin files have been modified to pull all nodes based on
the supplied hwaddr, and skip nodes that are disabled.
----------------------------------------------------------------------


git-svn-id: https://warewulf.lbl.gov/svn/trunk@1475 3fa55bd7-8392-0410-b022-a3fc164e7bfc